pyproject.toml
[tool.poetry]
name = "chaos-nightmare"
version = "0.1.0"
description = "crowd-sourced image generation for twitch"
authors = ["Robert Sivilli <robert.sivilli@gmail.com>"]
readme = "README.md"
packages = [{include = "chaos_nightmare"}]

[tool.poetry.dependencies]
python = "^3.10"
twitchio = "^2.6.0"
pydantic = "^2.1.1"
pydantic-settings = "^2.0.2"
python-json-logger = "^2.0.7"
pyvirtualcam = "^0.10.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pillow = "^10.0.0"
aiohttp = "^3.8.5"
asyncio = "^3.4.3"


[tool.poetry.group.Dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
